release: 0.15.0

This PR includes quite a few changes. Mostly around getting patch panels imported and connected to Circuits and Devices. Also added quite a few tests for various parts of importing and loading the adapter for Device42.

### Feature
* :zap: Add UUID attribute to all models for the Nautobot Data Target. ([`07c8a40`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/07c8a408efc1615536077f66680cd5a5b0dec267))

### Fix
* :bug: Tweak device import to ignore those where a Building can't be found. ([`a59fd82`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/a59fd82d18ed25a678455492e7134a9d8eac9cc4))
* :bug: Include VRF when trying to find IP Address for update ([`a164592`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/a164592a8dd959e5dc54027a887d84c026482338))
* :bug: Fix infinite loop ([`6c5d2e5`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/6c5d2e525b6d8a19ab38d1c9f8ba40739303d917))
* :bug: Tweak find_ipaddr and set_primary_from_dns methods to have IP Address returned once found. ([`b3f62db`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/b3f62db91262236fb4df18a656f103c52cfaa43a))
* :bug: Fix infinite loop when looking for IP in VRF ([`4d792eb`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/4d792eb573dfc65bbc3bad9054e3c0d675792cfe))
* :bug: Update IP Address delete to handle multiple results ([`e8478be`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/e8478bec62155ddc24b74c5e831df0a5a15d181d))
* :bug: Rewrite find_ipaddr to check VRFs ([`a6592b8`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/a6592b8613b6fce58090692364cd1720592d41b1))
* :bug: Subnet/IPAddress delete methods updated to look for vrf__name ([`d53badf`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/d53badf904cd4f4cf82d8e3f0fcc971507fed314))
* :bug: Add VRF as identifier to IP Address ([`20a1fa0`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/20a1fa0a0e247882688445c88afd12991e5fe3d5))
* :bug: Ensure all logging sends to message. ([`874db65`](https://github.com/networktocode-llc/cu-zest-ssot-device42/commit/874db650278d278f0b84fa933169dec40cc25658))

See merge request networktocode-llc/cu-zest-ssot-device42!45

+++ nautobot_ssot_device42/diffsync/from_d42/models/circuits.py
@@ -1,10 +1,11 @@
-"""DiffSyncModel DCIM subclasses for Nautobot Device42 data sync."""
+"""DiffSyncModel Circuit subclasses for Nautobot Device42 data sync."""
 
 from typing import List, Optional
 from uuid import UUID
 
 from diffsync import DiffSyncModel
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import ValidationError
 from django.utils.text import slugify
 from nautobot.circuits.models import Circuit as NautobotCircuit
 from nautobot.circuits.models import CircuitTermination as NautobotCT
@@ -13,7 +14,7 @@ from nautobot.dcim.models import Cable as NautobotCable
 from nautobot.dcim.models import Device as NautobotDevice
 from nautobot.dcim.models import Interface as NautobotInterface
 from nautobot.extras.models import Status as NautobotStatus
-from nautobot_ssot_device42.constant import INTF_SPEED_MAP
+from nautobot_ssot_device42.constant import INTF_SPEED_MAP, PLUGIN_CFG
 from nautobot_ssot_device42.utils import nautobot
 
 
@@ -52,8 +53,13 @@ class Provider(DiffSyncModel):
             if attrs.get("tags"):
                 for _tag in nautobot.get_tags(attrs["tags"]):
                     _provider.tags.add(_tag)
-            _provider.validated_save()
-            return super().create(ids=ids, diffsync=diffsync, attrs=attrs)
+            try:
+                _provider.validated_save()
+                return super().create(ids=ids, diffsync=diffsync, attrs=attrs)
+            except ValidationError as err:
+                if diffsync.job.debug:
+                    diffsync.job.log_warning(message=f"Unable to create {ids['name']} provider. {err}")
+                return None
 
     def update(self, attrs):
         """Update Provider object in Nautobot."""
@@ -75,13 +81,14 @@ class Provider(DiffSyncModel):
         """Delete Provider object from Nautobot.
 
         Because Provider has a direct relationship with Circuits it can't be deleted before them.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        self.diffsync.job.log_warning(message=f"Provider {self.name} will be deleted.")
-        super().delete()
-        provider = NautobotProvider.objects.get(id=self.uuid)
-        self.diffsync._objects_to_delete["provider"].append(provider)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            self.diffsync.job.log_warning(message=f"Provider {self.name} will be deleted.")
+            super().delete()
+            provider = NautobotProvider.objects.get(id=self.uuid)
+            self.diffsync.objects_to_delete["provider"].append(provider)  # pylint: disable=protected-access
         return self
 
 
@@ -157,7 +164,7 @@ class Circuit(DiffSyncModel):
         if attrs.get("type"):
             _circuit.type = nautobot.verify_circuit_type(attrs["type"])
         if attrs.get("status"):
-            _circuit.status = NautobotStatus.objects.get(name=self.get_circuit_status(attrs["status"]))
+            _circuit.status = NautobotStatus.objects.get(name=attrs["status"])
         if attrs.get("install_date"):
             _circuit.install_date = attrs["install_date"]
         if attrs.get("bandwidth"):
@@ -214,11 +221,12 @@ class Circuit(DiffSyncModel):
         """Delete Provider object from Nautobot.
 
         Because Provider has a direct relationship with Circuits it can't be deleted before them.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        self.diffsync.job.log_warning(message=f"Circuit {self.circuit_id} will be deleted.")
-        super().delete()
-        circuit = NautobotCircuit.objects.get(id=self.uuid)
-        self.diffsync._objects_to_delete["circuit"].append(circuit)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            self.diffsync.job.log_warning(message=f"Circuit {self.circuit_id} will be deleted.")
+            super().delete()
+            circuit = NautobotCircuit.objects.get(id=self.uuid)
+            self.diffsync.objects_to_delete["circuit"].append(circuit)  # pylint: disable=protected-access
         return self
