+++ nautobot_ssot_device42/diffsync/from_d42/models/ipam.py
@@ -80,13 +80,15 @@ class VRFGroup(DiffSyncModel):
         """Delete VRF object from Nautobot.
 
         Because VRF has a direct relationship with many other objects it can't be deleted before anything else.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        self.diffsync.job.log_warning(message=f"VRF {self.name} will be deleted.")
-        super().delete()
-        vrf = NautobotVRF.objects.get(**self.get_identifiers())
-        self.diffsync._objects_to_delete["vrf"].append(vrf)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            vrf = NautobotVRF.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_warning(message=f"VRF {self.name} will be deleted.")
+            self.diffsync.objects_to_delete["vrf"].append(vrf)  # pylint: disable=protected-access
         return self
 
 
@@ -159,13 +161,15 @@ class Subnet(DiffSyncModel):
         """Delete Subnet object from Nautobot.
 
         Because Subnet has a direct relationship with many other objects it can't be deleted before anything else.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        self.diffsync.job.log_warning(message=f"Subnet {self.network} will be deleted.")
-        super().delete()
-        subnet = NautobotPrefix.objects.get(id=self.uuid)
-        self.diffsync._objects_to_delete["subnet"].append(subnet)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            subnet = NautobotPrefix.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_debug(message=f"Subnet {self.network} will be deleted.")
+            self.diffsync.objects_to_delete["subnet"].append(subnet)  # pylint: disable=protected-access
         return self
 
 
@@ -217,9 +221,10 @@ class IPAddress(DiffSyncModel):
                 _ip.assigned_object_type = ContentType.objects.get(app_label="dcim", model="interface")
                 _ip.assigned_object_id = intf.id
             except NautobotInterface.DoesNotExist as err:
-                diffsync.job.log_debug(
-                    message=f"Unable to find Interface {attrs['interface']} for {attrs['device']}. {err}"
-                )
+                if diffsync.job.debug:
+                    diffsync.job.log_debug(
+                        message=f"Unable to find Interface {attrs['interface']} for {attrs['device']}. {err}",
+                    )
         if attrs.get("interface"):
             if re.search(r"[Ll]oopback", attrs["interface"]):
                 _ip.role = "loopback"
@@ -261,8 +266,14 @@ class IPAddress(DiffSyncModel):
                         _ip.validated_save()
                         nautobot.set_primary_ip_and_mgmt(_ip, _dev, _intf)
             except NautobotDevice.DoesNotExist:
+                if diffsync.job.debug:
+                    diffsync.job.log_debug(message=f"Unable to find Device {attrs['device']} for {_ip.address}.")
                 pass
             except NautobotInterface.DoesNotExist:
+                if diffsync.job.debug:
+                    diffsync.job.log_debug(
+                        message=f"Unable to find Interface {attrs['interface']} for device {attrs['device']} for {_ip.address}."
+                    )
                 pass
         return super().create(ids=ids, diffsync=diffsync, attrs=attrs)
 
@@ -271,7 +282,10 @@ class IPAddress(DiffSyncModel):
         try:
             _ipaddr = NautobotIPAddress.objects.get(id=self.uuid)
         except NautobotIPAddress.DoesNotExist:
-            print("IP Address passed to update but can't be found. This shouldn't happen. Why is this happening?!?!")
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_debug(
+                    message="IP Address passed to update but can't be found. This shouldn't happen. Why is this happening?!?!"
+                )
             return
         if attrs.get("available"):
             _ipaddr.status = (
@@ -288,9 +302,10 @@ class IPAddress(DiffSyncModel):
                 _ipaddr.assigned_object_type = ContentType.objects.get(app_label="dcim", model="interface")
                 _ipaddr.assigned_object_id = intf.id
             except NautobotInterface.DoesNotExist as err:
-                self.diffsync.job.log_debug(
-                    f"Unable to find Interface {attrs['interface']} for {attrs['device']}. {err}"
-                )
+                if self.diffsync.job.debug:
+                    self.diffsync.job.log_debug(
+                        message=f"Unable to find Interface {attrs['interface']} for {attrs['device']}. {err}"
+                    )
         elif attrs.get("device") and attrs["device"] == "":
             try:
                 intf = NautobotInterface.objects.get(device=_ipaddr.assigned_object.device, name=self.interface)
@@ -304,18 +319,20 @@ class IPAddress(DiffSyncModel):
                     _dev.primary_ip6 = None
                 _dev.validated_save()
             except NautobotInterface.DoesNotExist as err:
-                self.diffsync.job.log_debug(
-                    f"Unable to find Interface {attrs['interface']} for {str(_ipaddr.assigned_object.device)} {err}"
-                )
+                if self.diffsync.job.debug:
+                    self.diffsync.job.log_debug(
+                        message=f"Unable to find Interface {attrs['interface']} for {str(_ipaddr.assigned_object.device)} {err}"
+                    )
         elif attrs.get("interface") and attrs["interface"] == "":
             try:
                 intf = NautobotInterface.objects.get(name=self.interface, device__name=attrs["device"])
                 _ipaddr.assigned_object_type = ContentType.objects.get(app_label="dcim", model="interface")
                 _ipaddr.assigned_object_id = intf.id
             except NautobotInterface.DoesNotExist as err:
-                self.diffsync.job.log_debug(
-                    message=f"Unable to find Interface {self.interface} for {attrs['device']}. {err}"
-                )
+                if self.diffsync.job.debug:
+                    self.diffsync.job.log_debug(
+                        message=f"Unable to find Interface {self.interface} for {attrs['device']}. {err}"
+                    )
         if attrs.get("primary") and attrs["primary"] is not None:
             _device, _intf = False, False
             if attrs.get("device") and self.device != "":
@@ -352,13 +369,15 @@ class IPAddress(DiffSyncModel):
         """Delete IPAddress object from Nautobot.
 
         Because IPAddress has a direct relationship with many other objects it can't be deleted before anything else.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        print(f"IP Address {self.address} will be deleted. {self}")
-        super().delete()
-        ipaddr = NautobotIPAddress.objects.get(id=self.uuid)
-        self.diffsync._objects_to_delete["ipaddr"].append(ipaddr)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            ipaddr = NautobotIPAddress.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_debug(message=f"IP Address {self.address} will be deleted. {self}")
+            self.diffsync.objects_to_delete["ipaddr"].append(ipaddr)  # pylint: disable=protected-access
         return self
 
 
@@ -389,8 +408,8 @@ class VLAN(DiffSyncModel):
             try:
                 _site = NautobotSite.objects.get(name=ids["building"])
             except NautobotSite.DoesNotExist as err:
-                if PLUGIN_CFG.get("verbose_debug"):
-                    diffsync.job.log_warning(message=f"Unable to find Site {ids['building']}. {err}")
+                if diffsync.job.debug:
+                    diffsync.job.log_debug(message=f"Unable to find Site {ids['building']}. {err}")
         try:
             _vlan = NautobotVLAN.objects.get(name=ids["name"], vid=ids["vlan_id"], site=_site)
         except NautobotVLAN.DoesNotExist:
@@ -415,7 +434,8 @@ class VLAN(DiffSyncModel):
         try:
             _vlan.validated_save()
         except ObjectAlreadyExists as err:
-            diffsync.job.log_debug(message=f"{ids['name']} already exists. {err}")
+            if diffsync.job.debug:
+                diffsync.job.log_debug(message=f"{ids['name']} already exists. {err}")
         return super().create(ids=ids, diffsync=diffsync, attrs=attrs)
 
     def update(self, attrs):
@@ -440,11 +460,13 @@ class VLAN(DiffSyncModel):
         """Delete VLAN object from Nautobot.
 
         Because VLAN has a direct relationship with many other objects it can't be deleted before anything else.
-        The self.diffsync._objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
+        The self.diffsync.objects_to_delete dictionary stores all objects for deletion and removes them from Nautobot
         in the correct order. This is used in the Nautobot adapter sync_complete function.
         """
-        self.diffsync.job.log_warning(message=f"VLAN {self.name} {self.vlan_id} {self.building} will be deleted.")
-        super().delete()
-        vlan = NautobotVLAN.objects.get(id=self.uuid)
-        self.diffsync._objects_to_delete["vlan"].append(vlan)  # pylint: disable=protected-access
+        if PLUGIN_CFG.get("delete_on_sync"):
+            super().delete()
+            vlan = NautobotVLAN.objects.get(id=self.uuid)
+            if self.diffsync.job.debug:
+                self.diffsync.job.log_debug(message=f"VLAN {self.name} {self.vlan_id} {self.building} will be deleted.")
+            self.diffsync.objects_to_delete["vlan"].append(vlan)  # pylint: disable=protected-access
         return self
